+++ src/calliope/backend/parsing.py
@@ -46,6 +46,16 @@ class UnparsedConstraintDict(TypedDict):
     slices: NotRequired[dict[str, list[UnparsedEquationDict]]]
 
 
+class UnparsedPiecewiseConstraintDict(TypedDict):
+    description: NotRequired[str]
+    foreach: NotRequired[list]
+    where: NotRequired[str]
+    x_expression: Required[str]
+    x_values: Required[str]
+    y_expression: Required[str]
+    y_values: Required[str]
+
+
 class UnparsedExpressionDict(UnparsedConstraintDict):
     unit: NotRequired[str]
 
@@ -78,6 +88,7 @@ UNPARSED_DICTS = Union[
     UnparsedVariableDict,
     UnparsedExpressionDict,
     UnparsedObjectiveDict,
+    UnparsedPiecewiseConstraintDict,
 ]
 T = TypeVar("T", bound=UNPARSED_DICTS)
 
@@ -329,7 +340,7 @@ class ParsedBackendEquation:
     @overload
     def evaluate_expression(
         self,
-        backend_interface: backend_model.BackendModel,
+        backend_interface: backend_model.BackendModelGenerator,
         *,
         return_type: Literal["array"] = "array",
         references: Optional[set] = None,
@@ -340,7 +351,7 @@ class ParsedBackendEquation:
     @overload
     def evaluate_expression(
         self,
-        backend_interface: backend_model.BackendModel,
+        backend_interface: backend_model.BackendModelGenerator,
         *,
         return_type: Literal["math_string"],
         references: Optional[set] = None,
@@ -349,7 +360,7 @@ class ParsedBackendEquation:
 
     def evaluate_expression(
         self,
-        backend_interface: backend_model.BackendModel,
+        backend_interface: backend_model.BackendModelGenerator,
         *,
         return_type: Literal["array", "math_string"] = "array",
         references: Optional[set] = None,
@@ -439,11 +450,18 @@ class ParsedBackendComponent(ParsedBackendEquation):
         "global_expressions": expression_parser.generate_arithmetic_parser,
         "objectives": expression_parser.generate_arithmetic_parser,
         "variables": lambda x: None,
+        "piecewise_constraints": expression_parser.generate_arithmetic_parser,
     }
 
     def __init__(
         self,
-        group: Literal["variables", "global_expressions", "constraints", "objectives"],
+        group: Literal[
+            "variables",
+            "global_expressions",
+            "constraints",
+            "piecewise_constraints",
+            "objectives",
+        ],
         name: str,
         unparsed_data: T,
     ) -> None: